Skip AppTest if en_US is not the default locale

+++ taskcoach/tests/unittests/AppTest.py
@@ -41,17 +41,25 @@ class AppTests(test.TestCase):
         self.options = DummyOptions()
 
     def testAppProperties(self):
-        # Normally I prefer one assert per test, but creating the app is
-        # expensive, so we do all the queries in one test method.
-        app = application.Application(loadSettings=False, loadTaskFile=False)
-        wxApp = wx.GetApp()
-        self.assertEqual(meta.name, wxApp.GetAppName())
-        self.assertEqual(meta.author, wxApp.GetVendorName())
-        app.mainwindow._idleController.stop()
-        app.quitApplication()
-        app.mainwindow.Destroy()
-        application.Application.deleteInstance()
-        
+        import locale
+        if locale.getdefaultlocale()[0] != 'en_US':
+            # Somehow wx displays an error dialog box if en_US is not installed, when
+            # quitApplication() calls ProcessIdle and I don't know how to get rid of it.
+            # I don't know how to find out if en_US is installed either, so skip if
+            # it's not the default.
+            self.skipTest('Locale is not en_US')
+        else:
+            # Normally I prefer one assert per test, but creating the app is
+            # expensive, so we do all the queries in one test method.
+            app = application.Application(loadSettings=False, loadTaskFile=False)
+            wxApp = wx.GetApp()
+            self.assertEqual(meta.name, wxApp.GetAppName())
+            self.assertEqual(meta.author, wxApp.GetVendorName())
+            app.mainwindow._idleController.stop()
+            app.quitApplication()
+            app.mainwindow.Destroy()
+            application.Application.deleteInstance()
+
     def assertLanguage(self, expectedLanguage, locale=None):
         args = [self.options, self.settings]
         if locale:
